pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oncomatch-ai"
version = "0.1.0"
description = "Clinical Trial Matching System for Oncology Patients"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "OncoMatch AI Team"}
]
dependencies = [
    "pydantic>=2.7",
    "pandas>=2.2",
    "tenacity>=8.2",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "anthropic>=0.25",
    "google-generativeai>=0.5",
    "diskcache>=5.6",
    "scikit-learn>=1.4",
    "biomcp-python>=0.6",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "pre-commit>=3.7",
]

[project.scripts]
oncomatch = "oncomatch.match:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["oncomatch", "oncomatch.evaluation"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]   # errors, pyflakes, import sort
fixable = ["ALL"]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-q --tb=short"

